Move PhoneNumberContent component to a standalone file
