+++ js/src/components/contact-information/phone-number-card/phone-number-card.js
@@ -1,25 +1,18 @@
 /**
  * External dependencies
  */
-import {
-	getCountryCallingCode,
-	parsePhoneNumberFromString as parsePhoneNumber,
-} from 'libphonenumber-js';
 import { __ } from '@wordpress/i18n';
 import { useState, useEffect } from '@wordpress/element';
-import { Flex, FlexItem, FlexBlock, CardDivider } from '@wordpress/components';
+import { CardDivider } from '@wordpress/components';
 import { Spinner } from '@woocommerce/components';
 
 /**
  * Internal dependencies
  */
-import useCountryCallingCodeOptions from '.~/hooks/useCountryCallingCodeOptions';
-import Section from '.~/wcdl/section';
-import SelectControl from '.~/wcdl/select-control';
-import AppInputControl from '.~/components/app-input-control';
 import AccountCard, { APPEARANCE } from '.~/components/account-card';
 import AppButton from '.~/components/app-button';
 import AppSpinner from '.~/components/app-spinner';
+import EditPhoneNumberContent from './edit-phone-number-content';
 import './phone-number-card.scss';
 
 const noop = () => {};
@@ -31,107 +24,13 @@ const basePhoneNumberCardProps = {
 
 /**
  * @typedef { import(".~/hooks/useGoogleMCPhoneNumber").PhoneNumber } PhoneNumber
- * @typedef { import(".~/hooks/useGoogleMCPhoneNumber").PhoneNumberData } PhoneNumberData
+ * @typedef { import("./edit-phone-number-content").onPhoneNumberChange } onPhoneNumberChange
  */
-
-/**
- * @typedef {Object} ExtraPhoneNumberData
- * @property {boolean} isDirty Whether the phone number data contain unsaved changes.
- *
- * @typedef {PhoneNumberData & ExtraPhoneNumberData} CallbackPhoneNumberData
- */
-
-/**
- * @callback onPhoneNumberChange
- * @param {CallbackPhoneNumberData} phoneNumberData The changed phone number data.
- */
-
-function PhoneNumberContent( {
-	initCountry,
-	initNationalNumber,
-	onPhoneNumberChange,
-} ) {
-	const countryCallingCodeOptions = useCountryCallingCodeOptions();
-	const [ country, setCountry ] = useState( initCountry );
-	const [ number, setNumber ] = useState( initNationalNumber );
-
-	const handleChange = ( nextCountry, nextNumber ) => {
-		setCountry( nextCountry );
-		setNumber( nextNumber );
-
-		const parsed = parsePhoneNumber( nextNumber, nextCountry );
-		const isValid = parsed ? parsed.isValid() : false;
-
-		if ( parsed ) {
-			const isDirty =
-				nextCountry !== initCountry ||
-				parsed.nationalNumber !== initNationalNumber;
-
-			onPhoneNumberChange( {
-				...parsed,
-				isValid,
-				isDirty,
-			} );
-		} else {
-			const isDirty =
-				nextCountry !== initCountry ||
-				nextNumber !== initNationalNumber;
-
-			const countryCallingCode = nextCountry
-				? getCountryCallingCode( nextCountry )
-				: '';
-
-			onPhoneNumberChange( {
-				isValid,
-				isDirty,
-				countryCallingCode,
-				nationalNumber: nextNumber,
-				country: nextCountry,
-			} );
-		}
-	};
-
-	const handleCountryChange = ( nextCountry ) =>
-		handleChange( nextCountry, number );
-
-	const handleNumberChange = ( nextNumber ) =>
-		handleChange( country, nextNumber );
-
-	return (
-		<Section.Card.Body>
-			<Flex gap={ 4 }>
-				<FlexItem>
-					<SelectControl
-						label={ __(
-							'Country code',
-							'google-listings-and-ads'
-						) }
-						isSearchable
-						excludeSelectedOptions={ false }
-						options={ countryCallingCodeOptions }
-						selected={ country }
-						onChange={ handleCountryChange }
-					/>
-				</FlexItem>
-				<FlexBlock>
-					<AppInputControl
-						label={ __(
-							'Phone number',
-							'google-listings-and-ads'
-						) }
-						value={ number }
-						onChange={ handleNumberChange }
-					/>
-				</FlexBlock>
-			</Flex>
-		</Section.Card.Body>
-	);
-}
 
 function EditPhoneNumberCard( { phoneNumber, onPhoneNumberChange } ) {
 	const { loaded, data } = phoneNumber;
-	const phoneNumberContent = loaded ? (
-		<PhoneNumberContent
+	const cardContent = loaded ? (
+		<EditPhoneNumberContent
 			initCountry={ data.country }
 			initNationalNumber={ data.nationalNumber }
 			onPhoneNumberChange={ onPhoneNumberChange }
@@ -149,7 +48,7 @@ function EditPhoneNumberCard( { phoneNumber, onPhoneNumberChange } ) {
 			) }
 		>
 			<CardDivider />
-			{ phoneNumberContent }
+			{ cardContent }
 		</AccountCard>
 	);
 }
